+++ controls/models.py
@@ -9,7 +9,6 @@ from .oscal import Catalogs, Catalog
 import uuid
 import tools.diff_match_patch.python3 as dmp_module
 
-
 BASELINE_PATH = os.path.join(os.path.dirname(__file__),'data','baselines')
 
 class Statement(models.Model):
@@ -60,7 +59,7 @@ class Statement(models.Model):
         """Return the control content from the catalog"""
         # Get instance of the control catalog
         catalog = Catalog.GetInstance(catalog_key=self.sid_class)
-        catalog_control_dict = catalog.get_flattended_controls_all_as_dict()
+        catalog_control_dict = catalog.get_flattened_controls_all_as_dict()
         # Look up control by ID
         return catalog_control_dict[self.sid]