+++ siteapp/tests.py
@@ -16,6 +16,7 @@ import re
 from unittest import skip
 
 from django.conf import settings
+from django.contrib.auth.models import Permission
 from django.contrib.staticfiles.testing import StaticLiveServerTestCase
 from django.utils.crypto import get_random_string
 
@@ -238,6 +239,7 @@ class OrganizationSiteFunctionalTests(SeleniumTest):
         self.user.set_password(self.user.clear_password)
         self.user.save()
         self.user.reset_api_keys()
+        self.user.user_permissions.add(Permission.objects.get(codename='view_appsource'))
         self.client.login(username=self.user.username, password=self.user.clear_password)
 
         # Create a Portfolio and Grant Access
@@ -259,6 +261,7 @@ class OrganizationSiteFunctionalTests(SeleniumTest):
         self.user2.set_password(self.user2.clear_password)
         self.user2.save()
         self.user2.reset_api_keys()
+        self.user2.user_permissions.add(Permission.objects.get(codename='view_appsource'))
         self.client.login(username=self.user2.username, password=self.user2.clear_password)
         portfolio = Portfolio.objects.create(title=self.user2.username)
         portfolio.assign_owner_permissions(self.user2)
@@ -271,6 +274,7 @@ class OrganizationSiteFunctionalTests(SeleniumTest):
         self.user3.set_password(self.user3.clear_password)
         self.user3.save()
         self.user3.reset_api_keys()
+        self.user3.user_permissions.add(Permission.objects.get(codename='view_appsource'))
         self.client.login(username=self.user3.username, password=self.user3.clear_password)
         portfolio = Portfolio.objects.create(title=self.user3.username)
         portfolio.assign_owner_permissions(self.user3)
@@ -366,7 +370,7 @@ class GeneralTests(OrganizationSiteFunctionalTests):
         self.assertRegex(self.browser.title, "Welcome to Compliance Automation")
 
     def test_login(self):
-        # Test that a wrong password doesn't log us in.
+        # Test that a wrong pwd doesn't log us in.
         self._login(password=get_random_string(4))
         self.assertInNodeText("The username and/or password you specified are not correct.", "form#login_form .alert-danger")
 
@@ -813,7 +817,7 @@ class QuestionsTests(OrganizationSiteFunctionalTests):
         var_sleep(.5)
         self._test_api_get(["question_types_text", "q_text_with_default"], "I am a kiwi.")
 
-        # password-type question input (this is not a user password)
+        # password-type question input (this is not a user pwd)
         self.assertRegex(self.browser.title, "Next Question: password")
         self.fill_field("#inputctrl", "th1s1z@p@ssw0rd!")
         self.click_element("#save-button")
@@ -1224,6 +1228,9 @@ class OrganizationSettingsTests(OrganizationSiteFunctionalTests):
         var_sleep(0.5)
 
     def test_settings_page(self):
+        # Log in
+        var_sleep(.5)
+        self._login()
         # test navigating to settings page not logged in
         self.browser.get(self.url("/settings"))
         self.assertRegex(self.browser.title, "GovReady-Q")
@@ -1231,6 +1238,7 @@ class OrganizationSettingsTests(OrganizationSiteFunctionalTests):
         var_sleep(0.5)
 
         # login as user without admin privileges and test settings page unreachable
+        self.browser.get(self.url("/accounts/logout/"))
         self._login(self.user2.username, self.user2.clear_password)
         self.browser.get(self.url("/projects"))
         var_sleep(1)
@@ -1249,7 +1257,6 @@ class OrganizationSettingsTests(OrganizationSiteFunctionalTests):
 
         print("self.user is '{}'".format(self.user))
         print("self.user.username is '{}'".format(self.user.username))
-        # print("self.user.clear_password is '{}'".format(self.user.clear_password))
         print("self.user2.username is '{}'".format(self.user2.username))
 
         # SAMPLE NAVIGATING AND TESTING